Convert gensrcs to java_library to deprecate depfile
Cleanup usage of depfile in gensrcs to deprecate the property (aosp/2120835). The concept of depfile isn't transferable to Bazel, which requires listing the deps explicitly. The end goal of this change is to ensure all remaining gensrcs modules are convertible to Bazel by bp2build.

Test: m service-bluetooth
Bug: 179452413
Change-Id: I58d45d26ba4c9205e8020810319724479f21285e
Ignore-AOSP-First: The fixed modules don't exist in AOSP