Bluetooth timeout feature [1/2]
Add a bluetooth timeout setting and listen to changes to it as well as
the bluetooth radio and connection state. Schedule an alarm to turn bluetooth
off if it is enabled but not connected

Part 2 is in Settings.

Issue: calyxos#124
Change-Id: I4f4646139cd42b7ad20b486ed77dd00499aa3c6e
(cherry picked from commit 28da7976f4643c1b9c49edcefc8f620aeaa06ae2)

Issue: calyxos#830
Change-Id: I5133909a030b5bb940707320420a26d8e1eb4852

Alarm listener must not be a local variable

Issue: calyxos#1050
Change-Id: Ibb00c13ffe0fa5811c4fbd451acb8fa25fa58813

Use AOSP Settings for BLUETOOTH_OFF_TIMEOUT

This APEX module is not able to access LineageSettings, so the setting
is moved back to AOSP.

Change-Id: I887cc3cea6c7f59523680d1141f91727086e143e